rerun script 06
I haven't had luck re-creating the stat output errors, so I re-ran all of script 06 (creating new igraphs) and then checked the re-generated stats output files.
I found that the stat files from the aggregate networks *have the same names* as the corresponding stat files from the baseline networks (because they have the same filters, edge types, and are run at the state or port level). I was finally able to replicate the original stat file output by (1) creating new, unique names for aggregate network stat files, and (2) comparing the stats from the aggregate networks with the stats in the original files that appear to be from the baseline networks, but are actually aggregate network data.