improve label positions in plots
use a rescaling function to improve the position of node labels in network plots.

instead of editing the existing function (plot_comparable_networks), I created a second version of the function. new edits can be incorporated into the original version of the function once we're sure that it's working correctly.