fix script 01
In certain years, the statistics from the WA Coast port group were being saved into the data frame twice.
This was because of a misplaced `}` to close an if-else statement, so that if a port group didn't have a network in a given year (e.g., other Washington ports), the previous port group's stats (e.g., WA Coast) were re-saved into the output data frame.